# fix: update google/genai to v1.9.0 and switch to parametersJsonSchema for MCP related tools