## TLDR

google/genai v1.9.0 had introduced a behavior change:
- Before: `parameters` field from the returns of McpTool.
- After:  populating `parametersJsonSchema`

The goal of this change is only to move MCP related tool path to use the `parametersJsonSchema`, will leave all other FunctionDeclaration as it is,.

## Dive Deeper

## Reviewer Test Plan

<!-- when a person reviews your code they should ideally be pulling and running that code. How would they validate your change works and if relevant what are some good classes of example prompts and ways they can exercise your changes -->

Test steps:

```
npm pack 
RELEASE_TARBALL=$(pwd)/google-gemini-cli-0.1.12.tgz
ls -lh $RELEASE_TARBALL
npm install $RELEASE_TARBALL
./node_modules/.bin/gemini
```

The config I used to test the local build:
```
{
  "selectedAuthType": "gemini-api-key",
  "theme": "Dracula",
  "mcpServers": {
    "context7": {
      "command": "npx",
      "args": ["-y", "@upstash/context7-mcp"]
    },
    "git": {
      "command": "uvx",
      "args": [
        "mcp-server-git",
        "--repository",
        "/usr/local/google/home/wanlindu/gemini_cli_test"
      ]
    },
    "playwright": {
      "command": "npx",
      "args": ["@playwright/mcp@latest"]
    },
    "sequential-thinking": {
      "command": "npx",
      "args": ["-y","@modelcontextprotocol/server-sequential-thinking"]
    },
    "memory": {
      "command": "npx",
      "args": ["-y", "@modelcontextprotocol/server-memory"]
    },
    "time": {
      "command": "uvx",
      "args": ["mcp-server-time"]
    },
    "everything": {
      "command": "npx",
      "args": [
        "-y",
        "@modelcontextprotocol/server-everything"
      ]
    }
  },
  "preferredEditor": "nano"
}

```

All tool are ready:
```
 Configured MCP servers:

  🟢 context7 - Ready (2 tools)
    - context7__get-library-docs
    - context7__resolve-library-id

  🟢 git - Ready (13 tools)
    - git__git_add
    - git__git_branch
    - git__git_checkout
    - git__git_commit
    - git__git_create_branch
    - git__git_diff
    - git__git_diff_staged
    - git__git_diff_unstaged
    - git__git_init
    - git__git_log
    - git__git_reset
    - git__git_show
    - git__git_status

  🟢 playwright - Ready (25 tools)
    - playwright__browser_click
    - playwright__browser_close
    - playwright__browser_console_messages
    - playwright__browser_drag
    - playwright__browser_file_upload
    - playwright__browser_generate_playwright_test
    - playwright__browser_handle_dialog
    - playwright__browser_hover
    - playwright__browser_install
    - playwright__browser_navigate
    - playwright__browser_navigate_back
    - playwright__browser_navigate_forward
    - playwright__browser_network_requests
    - playwright__browser_pdf_save
    - playwright__browser_press_key
    - playwright__browser_resize
    - playwright__browser_select_option
    - playwright__browser_snapshot
    - playwright__browser_tab_close
    - playwright__browser_tab_list
    - playwright__browser_tab_new
    - playwright__browser_tab_select
    - playwright__browser_take_screenshot
    - playwright__browser_type
    - playwright__browser_wait_for

  🟢 sequential-thinking - Ready (1 tools)
    - sequential-thinking__sequentialthinking

  🟢 memory - Ready (9 tools)
    - memory__add_observations
    - memory__create_entities
    - memory__create_relations
    - memory__delete_entities
    - memory__delete_observations
    - memory__delete_relations
    - memory__open_nodes
    - memory__read_graph
    - memory__search_nodes

  🟢 time - Ready (2 tools)
    - time__convert_time
    - time__get_current_time

  🟢 everything - Ready (8 tools)
    - everything__add
    - everything__annotatedMessage
    - everything__echo
    - everything__getResourceReference
    - everything__getTinyImage
    - everything__longRunningOperation
    - everything__printEnv
    - everything__sampleLLM
   ```

Was able to conduct FunctionCalls using sequentialthinking and add(everything), seems like all tool calls are successful.
```
                                                                                                                                                                                                                                                   │
│  Agent powering down. Goodbye!                                                                                                                                                                                                                    │
│                                                                                                                                                                                                                                                   │
│  Interaction Summary                                                                                                                                                                                                                              │
│  Tool Calls:                 11 ( ✔ 11 ✖ 0 )                                                                                                                                                                                                      │
│  Success Rate:               100.0%                                                                                                                                                                                                               │
│  User Agreement:             100.0% (4 reviewed)                                                                                                                                                                                                  │
│                                                                                                                                                                                                                                                   │
│  Performance                                                                                                                                                                                                                                      │
│  Wall Time:                  9m 24s                                                                                                                                                                                                               │
│  Agent Active:               2m 19s                                                                                                                                                                                                               │
│    » API Time:               1m 56s (83.4%)                                                                                                                                                                                                       │
│    » Tool Time:              23.2s (16.6%)                                                                                                                                                                                                        │
│                                                                                                                                                                                                                                                   │
│                                                                                                                                                                                                                                                   │
│  Model Usage                  Reqs   Input Tokens  Output Tokens                                                                                                                                                                                  │
│  ───────────────────────────────────────────────────────────────                                                                                                                                                                                  │
│  gemini-2.5-pro                 15        234,591          5,278                                                                                                                                                                                  │
│                                                                                                                                                                                                                                                   │
│  Savings Highlight: 171,665 (73.2%) of input tokens were served from the cache, reducing costs.                                                                                                                                                   │
│                                                                                                                                                                                                                                                   │
│  » Tip: For a full token breakdown, run `/stats model`.                                                                                                                                                                                           │
│
```

## Testing Matrix

<!-- Before submitting please validate your changes on as many of these options as possible -->

|          | 🍏  | 🪟  | 🐧  |
| -------- | --- | --- | --- |
| npm run  | ❓  | ❓  |  ✅   |
| npx      | ❓  | ❓  | ❓  |
| Docker   | ❓  | ❓  | ❓  |
| Podman   | ❓  | -   | -   |
| Seatbelt | ❓  | -   | -   |

## Linked issues / bugs

Fixes #4301 
Fixes #1481