@iwakitakuma33 2025-07-15
related issue https://github.com/google-gemini/gemini-cli/issues/1481

Many users are unable to use mcp via gemini cli in this issue. Once you upgrade genai to version 1.9.0, the error on the genai side when using mcp will be resolved and you will be able to use mcp.