@iwakitakuma33 2025-07-16
@wanlin31 
Will this be released soon?

Does upgrading only v1.9.0 affect other features?

related issue closed. https://github.com/google-gemini/gemini-cli/pull/4196



I will close [PR](https://github.com/google-gemini/gemini-cli/pull/4253) if this PR is addressed.